// groceries/src/models/Category.ts
export enum Category {
    ALCOHOL = 'Alcohol',
    BAKERY = 'Bakery',
    DAIRY_N_EGGS = 'Dairy & Eggs',
    DRINKS = 'Drinks',
    FROZEN = 'Frozen',
    HOME_N_HEALTH = 'Home & Health',
    MEAT_FISH_N_PROTEIN = 'Meat, Fish & Protein',
    PANTRY = 'Pantry',
    PET_PRODUCTS = 'Pet Products',
    PREPARED = 'Prepared',
    PRODUCE = 'Produce',
    SNACKS = 'Snacks',
}

export default Category;